+++ website/components/SideNavItem/SideNavItem.tsx
@@ -28,7 +28,7 @@ export const SideNavItem = ({
         align="center"
         transition="all 0.2s ease"
         opacity={active ? 1 : 0.9}
-        color={active ? "blue.500" : color}
+        color={active ? "teal.400" : color}
         _hover={{
           cursor: "pointer",
           opacity: 1,