+++ website/containers/CodeHighlight/CodeHighlight.tsx
@@ -1,15 +1,19 @@
 import { useCardBg } from "@/hooks/useCardBg";
-import { Stack } from "@chakra-ui/react";
+import { Stack, StackProps } from "@chakra-ui/react";
 import { Suspense } from "react";
+import CopyButton from "../../components/CopyButton/CopyButton";
 import LazyCodeLoader from "../../components/LazyCodeLoader/LazyCodeLoader";
 import { CodeExample } from "../../mdx/server";
 
-type CodeHighlightProps = {
+type CodeHighlightProps = StackProps & {
   code: CodeExample | undefined;
+  path?: string;
 };
 
-const CodeHighlight = ({ code }: CodeHighlightProps) => {
+const CodeHighlight = ({ code, path, ...rest }: CodeHighlightProps) => {
   const bg = useCardBg();
+
+  const { sx } = rest;
 
   if (!code) return null;
 
@@ -17,17 +21,22 @@ const CodeHighlight = ({ code }: CodeHighlightProps) => {
     <Stack
       sx={{
         width: "auto",
+        maxWidth: "100%",
         bg,
         mr: "auto",
         alignSelf: "flex-start",
         shadow: "sm",
         rounded: "md",
         overflow: "hidden",
+        position: "relative",
+        sx,
       }}
+      {...rest}
     >
       <Suspense fallback={<div>Loading...</div>}>
         <LazyCodeLoader code={code?.code} />
       </Suspense>
+      <CopyButton code={code?.code} position="absolute" top={4} right={4} />
     </Stack>
   );
 };