+++ website/mdx/server.tsx
@@ -1,27 +1,17 @@
-import * as chakraButtons from "@chakra-ui/button";
-import * as chakraColorMode from "@chakra-ui/color-mode";
-import * as chakraHooks from "@chakra-ui/hooks";
-import * as chakraLayout from "@chakra-ui/layout";
-import * as chakraTable from "@chakra-ui/table";
-import * as framerMotion from "framer-motion";
 import fs from "fs";
-import { bundleMDX } from "mdx-bundler";
 import path from "path";
-import * as reactIconsRI from "react-icons/ri";
 
 const EXAMPLES_PATH = path.join(process.cwd(), "code-samples/examples");
-const SNIPPETS_PATH = path.join(process.cwd(), "code-samples/snippets");
 const RECIPES_PATH = path.join(process.cwd(), "code-samples/recipes");
-const DOCS_PATH = path.join(process.cwd(), "docs");
 
 // recursively read in all files in the examples directory except for the index.ts file and output an array of strings containing the stringified buffer
-export const getFileStrings = (directory = EXAMPLES_PATH) => {
+export const getCodeExamples = (directory = EXAMPLES_PATH) => {
   let files: CodeExample[] = [];
   const filesInDirectory = fs.readdirSync(directory);
   for (const file of filesInDirectory) {
     const absolute = path.join(directory, file);
     if (fs.statSync(absolute).isDirectory()) {
-      files = files.concat(getFileStrings(absolute));
+      files = files.concat(getCodeExamples(absolute));
     } else {
       const filename = path.basename(absolute);
       const filepath = path.join(directory, filename);
@@ -38,7 +28,7 @@ export const getFileStrings = (directory = EXAMPLES_PATH) => {
   return files;
 };
 
-export const getFileString = (filepath: string): CodeExample | undefined => {
+export const getCodeExample = (filepath: string): CodeExample | undefined => {
   try {
     const absPath = path.join(process.cwd(), filepath);
     const file = fs.readFileSync(absPath);
@@ -55,7 +45,7 @@ export const getFileString = (filepath: string): CodeExample | undefined => {
 export type CodeExample = {
   code: string;
   filename: string;
-  filepath: string;
+  filepath?: string;
   language?: string;
 };
 
@@ -107,115 +97,3 @@ export function dirToFileArray(dir = RECIPES_PATH): MultipleCodeSampleResults {
 
   return results;
 }
-
-type Doc = {
-  code: string;
-  filename: string;
-};
-
-const chakra = Object.keys({
-  ...chakraHooks,
-  ...chakraLayout,
-  ...chakraButtons,
-  ...chakraColorMode,
-  ...chakraTable,
-}).filter((key) => key !== "__esModule");
-
-const framer = Object.keys(framerMotion).filter((key) => key !== "__esModule");
-
-const ri = Object.keys(reactIconsRI).filter((key) => key !== "__esModule");
-
-const SECTIONS_PATH = path.join(process.cwd(), "sections");
-
-export const getSourceOfFile = (filepath: string) => {
-  return fs.readFileSync(
-    path.join(SECTIONS_PATH, filepath, "index.mdx"),
-    "utf-8"
-  );
-};
-
-// export const getSections = (): {
-//   frontmatter: FrontMatter;
-//   slug: string;
-// }[] => {
-//   return fs.readdirSync(SECTIONS_PATH).map((filepath) => {
-//     const source = getSourceOfFile(filepath);
-//     const slug = filepath.replace(/\.mdx?$/, "");
-//     const { data } = matter(source);
-
-//     return {
-//       frontmatter: data as FrontMatter,
-//       slug: slug,
-//     };
-//   });
-// };
-
-export const getSection = async (slug: string) => {
-  const source = getSourceOfFile(slug);
-  const imagesUrl = `/img/blog/${slug}`;
-  const directory = path.join(SECTIONS_PATH, slug);
-
-  if (process.platform === "win32") {
-    process.env.ESBUILD_BINARY_PATH = path.join(
-      process.cwd(),
-      "node_modules",
-      "esbuild",
-      "esbuild.exe"
-    );
-  } else {
-    process.env.ESBUILD_BINARY_PATH = path.join(
-      process.cwd(),
-      "node_modules",
-      "esbuild",
-      "bin",
-      "esbuild"
-    );
-  }
-
-  const { code, frontmatter } = await bundleMDX({
-    source,
-    cwd: directory,
-    mdxOptions: (options) => {
-      options.remarkPlugins = [...(options.remarkPlugins ?? [])];
-      return options;
-    },
-    esbuildOptions: (options) => {
-      options.outdir = path.join(process.cwd(), "public", imagesUrl);
-      options.loader = {
-        ...options.loader,
-        ".webp": "file",
-        ".jpeg": "file",
-        ".jpg": "file",
-        ".svg": "file",
-        ".png": "file",
-        ".gif": "file",
-      };
-
-      options.publicPath = imagesUrl;
-      options.write = true;
-      return options;
-    },
-    globals: {
-      "@chakra-ui/react": {
-        varName: "chakra",
-        namedExports: chakra,
-        defaultExport: false,
-      },
-      "framer-motion": {
-        varName: "framer",
-        namedExports: framer,
-        defaultExport: false,
-      },
-      "@react-icons/ri": {
-        varName: "ri",
-        namedExports: ri,
-        defaultExport: false,
-      },
-    },
-  });
-
-  return {
-    frontmatter,
-    code,
-  };
-};